requirements.txt
CTkMenuBar==0.8
CTkMenuBar==0.8
CTkMessagebox==2.5
CTkMessagebox==2.5
customtkinter==5.2.2
Pillow==9.5.0
Pillow==10.3.0
pypdl==1.3.2
Requests==2.31.0
